requirements.txt
google-genai
python-dotenv
uagents
boto3